problem with connect in MainWindow

+++ libs/PhCommonUI/PhLeap.h
@@ -24,7 +24,7 @@ public:
 
 signals:
     void setPosition(int move);
-    void setRate(int move);
+    void setRate(float move);
 
 
 };
